raftstore: extract region::Task
