+++ components/tikv_misc/src/lib.rs
@@ -30,6 +30,7 @@ pub mod peer_storage;
 pub mod raftstore_bootstrap;
 pub mod raftstore_callback;
 pub mod region_snapshot;
+pub mod region_task;
 pub mod storage_key;
 pub mod store_info;
 pub mod store_util;
