+++ components/tikv_misc/src/region_task.rs
@@ -1,0 +1,60 @@
+use engine::Snapshot;
+use raft::eraftpb::Snapshot as RaftSnapshot;
+use std::fmt::{self, Display, Formatter};
+use std::sync::Arc;
+use std::sync::atomic::AtomicUsize;
+use std::sync::mpsc::SyncSender;
+use tikv_util::escape;
+
+/// Region related task
+#[derive(Debug)]
+pub enum Task {
+    Gen {
+        region_id: u64,
+        raft_snap: Snapshot,
+        kv_snap: Snapshot,
+        notifier: SyncSender<RaftSnapshot>,
+    },
+    Apply {
+        region_id: u64,
+        status: Arc<AtomicUsize>,
+    },
+    /// Destroy data between [start_key, end_key).
+    ///
+    /// The deletion may and may not succeed.
+    Destroy {
+        region_id: u64,
+        start_key: Vec<u8>,
+        end_key: Vec<u8>,
+    },
+}
+
+impl Task {
+    pub fn destroy(region_id: u64, start_key: Vec<u8>, end_key: Vec<u8>) -> Task {
+        Task::Destroy {
+            region_id,
+            start_key,
+            end_key,
+        }
+    }
+}
+
+impl Display for Task {
+    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
+        match *self {
+            Task::Gen { region_id, .. } => write!(f, "Snap gen for {}", region_id),
+            Task::Apply { region_id, .. } => write!(f, "Snap apply for {}", region_id),
+            Task::Destroy {
+                region_id,
+                ref start_key,
+                ref end_key,
+            } => write!(
+                f,
+                "Destroy {} [{}, {})",
+                region_id,
+                escape(start_key),
+                escape(end_key)
+            ),
+        }
+    }
+}
